FU-158 refactor: group photographer containers
