+++ src/app/photographer/layout.tsx
@@ -1,5 +1,5 @@
 import styles from "@/styles/page.module.css";
-import Header from "@/containers/main/header";
+import Header from "@/containers/photographer/main/header";
 
 const PhotographerLayout = ({
   children,
