+++ src/app/photographer/new-product/page.tsx
@@ -1,4 +1,4 @@
-import ProductForm from "@/containers/product/product-form";
+import ProductForm from "@/containers/photographer/product/product-form";
 
 const NewProduct = () => {
   return (
