+++ src/containers/photographer/product/header/product-icon.tsx
